Added new Poseidon implementation
